% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CopyNeutralIMA.R
\docType{package}
\name{CopyNeutralIMA-package}
\alias{CopyNeutralIMA}
\alias{CopyNeutralIMA-package}
\title{Reference samples for analysis of IlluminaHumanMethylation arrays.}
\description{
The CopyNeutralIMA package provides a set of 51 samples for
IlluminaHumanMethylation450k arrays and a set of 13 samples for IlluminaHumanMethylationEPIC arrays.
}
\details{
The package includes two RGChannelSetExtended objects, one with 51 samples hybridized with the IlluminaHumanMethylation450k array and another one with 13 samples hybridized with the IlluminaHumanMethylationEPIC array.
The provided samples consist of material from healthy patients with nominally no copy number aberations.
Users may use this data package to complement their copy number profiling analysis or for testing other tools.
}
\author{
\strong{Maintainer}: Xavier Pastor Hostench \email{x.pastorhostench@dkfz-heidelberg.de}

Authors:
\itemize{
  \item Moritz Przybilla \email{m.przybilla@dkfz-heidelberg.de}
}

}
