% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{package}
\name{CopyNeutralIMA}
\alias{CopyNeutralIMA}
\alias{CopyNeutralIMA-package}
\title{Reference samples for analysis of InfiniumHumanMethylation arrays.}
\format{RGChannelSetExtended wit 51 copy neutral samples.}
\source{
\url{https://www.ncbi.nlm.nih.gov/geo/}
}
\description{
The CopyNeutralIMA package provides a set of 51 samples for
IlluminaHumanMethylation450k arrays and 13 for IlluminaHumanMethylationEPIC arrays.
}
\details{
The package includes an RGChannelSetExtended object for each array type, with 51 samples hybridized with the InfiniumHumanMethylation450k array and 13 with the InfiniumHumanMethylationEPIC array.
The provided samples consist of material from healthy individuals with nominally no copy number aberrations.
Users may use this data package to complement their copy number profiling analysis or for testing other tools.
}
