% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{getCopyNeutralRGSet}
\alias{getCopyNeutralRGSet}
\title{Accessor to the data objects.}
\usage{
getCopyNeutralRGSet(ima)
}
\arguments{
\item{ima}{a character string identifying for which array type retrieve data. Valid values are 'IlluminaHumanMethylation450k' and 'IlluminaHumanMethylationEPIC'.}
}
\value{
A RGChannelSetExtended object
}
\description{
Accessor to the data objects.
}
\examples{
rgset_450k <- getCopyNeutralRGSet('IlluminaHumanMethylation450k')
rgset_450k
}
